cloud-functions-demo/function/requirements.txt
google-cloud-storage>=3.1.0